ath10k:  Retry peer delete.
Sometimes firmware looses a message (specifically, some CT firmware
does when it is forced into CE polling mode due to detecting
lack of WMI keep-alives from the driver).

Since we are waiting 3 seconds anyway, retry it 3 times
with one second delay.  This may give firmware the ability
to recover and stay better in sync with the driver.

Signed-off-by: Ben Greear <greearb@candelatech.com>